Nan int32

+++ runtime/Include/bigloo_int.h
@@ -3,7 +3,7 @@
 /*    -------------------------------------------------------------    */
 /*    Author      :  Manuel Serrano                                    */
 /*    Creation    :  Thu Mar  2 05:40:03 2017                          */
-/*    Last change :  Sat Apr 21 19:24:55 2018 (serrano)                */
+/*    Last change :  Sat Apr 21 20:21:46 2018 (serrano)                */
 /*    Copyright   :  2017-18 Manuel Serrano                            */
 /*    -------------------------------------------------------------    */
 /*    Bigloo INTEGERs                                                  */
@@ -45,15 +45,15 @@ extern "C" {
 #  define BGL_LONG_MAX (INT32_MAX)
 
 #  define BINT( i ) ((obj_t)(((long)((int32_t)i)) | TAG_INT))
-#  define CINT( i ) ((long)((int32_t)i))
+#  define CINT( i ) ((long)((int32_t)((long)i)))
 #  define ADDFX( x, y ) BINT( CINT( x ) + CINT( y ) )
 #  define SUBFX( x, y ) BINT( CINT( x ) - CINT( y ) )
 
-#  define LTFX( x, y ) ((int32_t)(x) < (int32_t)(y))
-#  define LEFX( x, y ) ((int32_t)(x) <= (int32_t)(y))
-#  define GTFX( x, y ) ((int32_t)(x) > (int32_t)(y))
-#  define GEFX( x, y ) ((int32_t)(x) >= (int32_t)(y))
-#  define EGFX( x, y ) ((x) == (y))
+#  define LTFX( x, y ) ((int32_t)((long)(x)) < (int32_t)((long)(y)))
+#  define LEFX( x, y ) ((int32_t)((long)(x)) <= (int32_t)((long)(y)))
+#  define GTFX( x, y ) ((int32_t)((long)(x)) > (int32_t)((long)(y)))
+#  define GEFX( x, y ) ((int32_t)((long)(x)) >= (int32_t)((long)(y)))
+#  define EGFX( x, y ) (((long)(x)) == ((long)(y)))
 #endif
 
 #define ODDP_FX( i )  ((i) & 1)
